requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --allow-unsafe --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.7.4
    # via signxml
cffi==1.16.0
    # via cryptography
cryptography==45.0.7
    # via
    #   -r requirements.in
    #   pyopenssl
    #   signxml
defusedxml==0.7.1
    # via -r requirements.in
django==4.2.24
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.0
    # via -r requirements.in
importlib-metadata==8.7.0
    # via -r requirements.in
jsonschema==4.25.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
lxml==5.4.0
    # via
    #   -r requirements.in
    #   signxml
marshmallow==3.26.1
    # via -r requirements.in
packaging==24.2
    # via marshmallow
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via -r requirements.in
pydantic-core==2.33.2
    # via pydantic
pyopenssl==25.1.0
    # via -r requirements.in
pytz==2025.2
    # via -r requirements.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
signxml==4.1.0
    # via -r requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.14.0
    # via
    #   -r requirements.in
    #   asgiref
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==80.9.0
    # via -r requirements.in
